playground/rtakashi/djangochannelsrestframework/requirements.txt
asgiref
django
django-otp
sqlparse
channels
channels_redis
daphne
djangorestframework
djangochannelsrestframework
